books-pipeline/requirements.txt
requests
beautifulsoup4
lxml
pandas
numpy
pyarrow
python-dotenv